# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Larynx word etymology and language API.",
        "title": "Larynx API",
        "contact": {},
        "version": "preview"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/geography/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a GeoJSON FeatureCollection for every Geography node whose name matches the id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "geography"
                ],
                "summary": "Get geography by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The geography name to look up",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.geographyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "server.feature": {
            "type": "object",
            "properties": {
                "geometry": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "properties": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "type": {
                    "type": "string",
                    "example": "Feature"
                }
            }
        },
        "server.geoJSON": {
            "type": "object",
            "properties": {
                "features": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/server.feature"
                    }
                },
                "type": {
                    "type": "string",
                    "example": "FeatureCollection"
                }
            }
        },
        "server.geographyResponse": {
            "type": "object",
            "properties": {
                "geojson": {
                    "$ref": "#/definitions/server.geoJSON"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AdminJWTAuth": {
            "description": "Admin JWT for blog write endpoints (create/update/delete)",
            "type": "apiKey",
            "name": "X-Admin-JWT",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Client bearer token. Enter the full value as: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}